Refactor generation of member expressions.
There were a few formatting issues with Expand-MemberExpression that
would have been more difficult to fix with the way it was implemented.
It now uses a few predefined strings as templates along with a helper
class.  The same issues were present in the way type expressions were
being generated.  That has been separated out into it's own class.

- Refactor Expand-MemberExpression to use string templates that can be
  customized using the exported variable PSESHLTemplates.

- Fix type expressions generating incorrectly for nonpublic generic
  types.

- Fix type expressions not correctly picking the best TypeAccelerators.

- Fix several formatting issues with Expand-MemberExpression.

- Add ability to select a template with a parameter.

- Add template using InvokeMember as default for methods and
  constructors. The GetMethod/Invoke template can be chosen by
  specifying the `VerboseInvokeMethod` template.

- Add priority to Properties over Fields and Methods with lower
  argument counts in Expand-MemberExpression.

- Expand comment based help.