Added PSEditorCommand attribute for functions (#9)
Functions defined with the PSEditorCommand attribute will get the
Context parameter added and be loaded as editor commands in
EditorServices when exported.  It can take the named parameters
Name, DisplayName, SuppressOutput, and SkipRegister. If SkipRegister
is true, the Context parameter will be added to the function but it
will not be loaded as a editor command.

The project is now buildable from VSCode tasks.  Previously it had to
be built from the integrated terminal directly due to parsing issues
around EditorServices classes not being loaded.

- Refactored internal methods of interacting with EditorServices and
  the workspace.

- Removed a shelved editor command that wasn't working as expected.