Use PSStringTemplate for code generation (#14)
* Use PSStringTemplate for code generation

Previously code generation was done using strings and the format
operator.  This made it difficult to maintain and add features. Code
generation now uses the PSStringTemplate module that utilizes the
Antlr4 StringTemplate engine which makes it a lot easier to make changes
to Expand-MemberExpression, and add more similar features.

Also added a temporary workaround for an issue where importing editor
commands could cause a null reference exception if a class within
this module had changed recently. This was because child scopes in
the module were still resolving the previous class.  To workaround
this, the psm1 file now retrieves the correct assemblies, and the
child scopes resolve from the RuntimeAssembly.GetType() method. This
will be better resolved in the future by porting the classes to C#.

- Expand-MemberExpression now works correctly for public members. For
  properties nothing will be expanded, but for methods it will generate
  parameter name comments.

- Added the NoParameterNameComments switch to Expand-MemberExpression.

- Fixed an issue with the build script around PSScriptAnalyzer and type
  resolution.