Fix position tracking in Set-ScriptExtent (#16)
* Fix position tracking in Set-ScriptExtent

This change enables inserting text at multiple positions without having
to track the updated position. This is more of a temporary solution
because it can't handle changes in multiple segments, and completely
falls apart on something like an undo.  To get around this, the extents
only live for the duration of Set-ScriptExtent and are recreated for every
subsequent run.

- Renamed Set-ExtentText to Set-ScriptExtent to match other extent related
  functions.

- Added a custom extent class that watches the current file for changes
  and adjusts the position accordingly.

- Added parameters 'AsString' and 'AsArray' to Set-ExtentText to allow
  easier creation of string expressions.

- Fixed an issue with ConvertTo-ScriptExtent where it couldn't create an
  extent from just an offset.