Refactor Context parameter injection (#17)
* Refactor Context parameter injection

This change cleans up the process Import-EditorCommand uses to inject
the Context parameter into functions. The original design attempted to
use some of the same processes used by the PowerShell engine for the
CmdletBinding attribute.

This process now manually creates a ParameterAst object with empty
extents to avoid issues with debugging.  The Context parameter now
also has a default value, so it will no longer be null if the function
is called outside of editor commands.

* Fix references to the Context parameter

- Remove the function TryGetEditorContext. This was a placeholder until
  the Context parameter could have a default value.

- Remove classes for generating parameter metadata.

- Fix references to the Context parameter.

* New line at EOF...