Add Set-RuleSuppression function (#18)
* Fix build issues missed from last change

* Add the Set-RuleSuppression function

This function creates and inserts working SuppressMessage attributes
targeting either the violating Ast itself, or the closest ParamBlock,
FunctionMember, or TypeDefinition depending on the Ast type.

If invoked as an editor command, it will suppress the Ast closest to
the cursor (if it has a violation).  You can also pipe Asts to it and
it will automatically determine if a suppression is needed and insert
accordingly.

You can use this to suppress entire files, or specific blocks (like
psake "Properties" blocks) with a single command.